+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# Next version
+
+Added
+  * AZ-802: Add timeouts option
+
 # v4.3.0 - 2022-03-18
 
 Added